// extensions/sdk/src/WebMessages.ts
/* This Source Code Form is subject to the terms of the Mozilla Public
 * License, v. 2.0. If a copy of the MPL was not distributed with this
 * file, You can obtain one at http://mozilla.org/MPL/2.0/. */

export enum WebMessages {
  WebCheck = "rally-sdk.web-check",  
  WebCheckResponse = "rally-sdk.web-check-response",

  CompleteSignup = "rally-sdk.complete-signup",
  CompleteSignupResponse = "rally-sdk.complete-signup-response",
  
  ChangeState = "rally-sdk.change-state"
}